requirements.txt
pyTelegramBotAPI
yt-dlp
google-generativeai
gTTS
pydub
